Add admin check to SplashActivity
This commit introduces an admin check in the SplashActivity.
If the user is an admin, a dialog with options to navigate to Login Activity or Main Activity will be shown.
Otherwise, the existing location check and city validation logic will be executed.

Additionally, the message for restricted access has been updated from "This app is not available in your current location." to "This app is not available in your city."
The exit dialog now uses `MaterialAlertDialogBuilder`.